+++ cpp/oneapi/dal/util/load_graph.hpp
@@ -17,7 +17,6 @@
 #pragma once
 
 #include <fstream>
-// #include <istream>
 
 #include "oneapi/dal/data/graph_common.hpp"
 #include "oneapi/dal/data/undirected_adjacency_array_graph.hpp"
@@ -25,8 +24,7 @@
 
 namespace oneapi::dal::preview::load_graph {
 
-template <typename Input = edge_list<int32_t>,
-          //   typename Allocator = std::allocator<empty_value>,
+template <typename Input  = edge_list<int32_t>,
           typename Output = undirected_adjacency_array_graph<empty_value,
                                                              empty_value,
                                                              empty_value,
@@ -46,19 +44,9 @@ edge_list<std::int32_t> load_edge_list(const std::string &name) {
     using int_t = std::int32_t;
 
     std::ifstream file(name);
-    std::string line;
-
-    int_t lines = 0;
-    while (std::getline(file, line)) {
-        lines++;
-    }
-    file.clear();
-    file.seekg(0, std::ios_base::beg);
+    edge_list<int_t> elist;
 
     char source_vertex[32], destination_vertex[32];
-
-    edge_list<int_t> elist;
-    elist.reserve(lines);
     while (file >> source_vertex >> destination_vertex) {
         auto edge = std::make_pair(daal::services::daal_string_to_int(&source_vertex[0], 0),
                                    daal::services::daal_string_to_int(&destination_vertex[0], 0));
